Renamed server.cc -> ctfo_server.cc, server.h -> ctfo_server.h
