+++ ctfo_server.cc
@@ -26,7 +26,7 @@ SOFTWARE.
 // Enable "$REST/data/user?export" unconditionally.
 #define CURRENT_ALLOW_STORAGE_EXPORT_FROM_MASTER
 
-#include "server.h"
+#include "ctfo_server.h"
 
 #include "../Current/Bricks/dflags/dflags.h"
